+++ datastore/load_test.go
@@ -15,6 +15,7 @@
 package datastore
 
 import (
+	"context"
 	"fmt"
 	"reflect"
 	"testing"
@@ -1226,9 +1227,81 @@ func TestLoadNull(t *testing.T) {
 	}
 }
 
-// 	var got2 struct{ S []Pet }
-// 	if err := LoadStruct(&got2, []Property{{Name: "S", Value: nil}}); err != nil {
-// 		t.Fatal(err)
-// 	}
+type KeyLoaderEnt struct {
+	A int
+	K *Key
+}
 
-// }
+func (e *KeyLoaderEnt) Load(p []Property) error {
+	e.A = 2
+	return nil
+}
+
+func (e *KeyLoaderEnt) LoadKey(k *Key) error {
+	e.K = k
+	return nil
+}
+
+func (e *KeyLoaderEnt) Save() ([]Property, error) {
+	return []Property{{Name: "A", Value: int64(3)}}, nil
+}
+
+func TestKeyLoaderEndToEnd(t *testing.T) {
+	keys := []*Key{
+		NameKey("testKind", "first", nil),
+		NameKey("testKind", "second", nil),
+	}
+
+	entity1 := &pb.Entity{
+		Key: keyToProto(keys[0]),
+		Properties: map[string]*pb.Value{
+			"A": {ValueType: &pb.Value_IntegerValue{IntegerValue: 1}},
+			"B": {ValueType: &pb.Value_StringValue{StringValue: "one"}},
+		},
+	}
+	entity2 := &pb.Entity{
+		Key: keyToProto(keys[1]),
+		Properties: map[string]*pb.Value{
+			"A": {ValueType: &pb.Value_IntegerValue{IntegerValue: 2}},
+			"B": {ValueType: &pb.Value_StringValue{StringValue: "two"}},
+		},
+	}
+
+	client, srv, cleanup := newMock(t)
+	defer cleanup()
+
+	srv.addRPC(&pb.LookupRequest{
+		ProjectId:  "projectID",
+		DatabaseId: "",
+		Keys: []*pb.Key{
+			keyToProto(keys[0]),
+			keyToProto(keys[1]),
+		},
+	},
+		&pb.LookupResponse{
+			Found: []*pb.EntityResult{
+				{
+					Entity:  entity1,
+					Version: 1,
+				},
+				{
+					Entity:  entity2,
+					Version: 1,
+				},
+			},
+		})
+
+	ctx := context.Background()
+
+	dst := make([]*KeyLoaderEnt, len(keys))
+	err := client.GetMulti(ctx, keys, dst)
+	if err != nil {
+		t.Fatalf("client.Get: %v", err)
+	}
+
+	for i := range dst {
+		if !testutil.Equal(dst[i].K, keys[i]) {
+			t.Fatalf("unexpected entity %d to have key %+v, got %+v", i, keys[i], dst[i].K)
+		}
+	}
+}
